+++ library/src/main/java/io/sfrei/tracksearch/utils/json/JsonElement.java
@@ -51,4 +51,8 @@ public class JsonElement extends JsonUtility {
         return jsonNode == null;
     }
 
+    public boolean notNull() {
+        return jsonNode != null;
+    }
+
 }