+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 0.5.4 - unreleased
 ------------------
 
+- Added optional path to YouTube tracks content
 - Fixed YouTube cipher field and added optional one
 
 0.5.3